+++ Code/CryEngine/CrySystem/Platform/Mac/platform_mac_files.cmake
@@ -8,8 +8,3 @@
 # remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #
-
-set(FILES
-    ../../SystemUtilsApple.h
-    ../../SystemUtilsApple.mm
-)
