+++ Code/CryEngine/CrySystem/crysystem_mac_files.cmake
@@ -9,7 +9,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #
 
-set(FILES
-    SystemUtilsApple.h
-    SystemUtilsApple.mm
-)
